@@ -1,11 +1,12 @@
 package com.polidea.multiplatformbleadapter;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import java.util.List;
 import java.util.UUID;
 
+/** @noinspection unused*/
 public class Device {
 
     private String id;
@@ -43,7 +44,7 @@ public class Device {
         return rssi;
     }
 
-    public void setRssi(Integer rssi) {
+    public void setRssi(@Nullable Integer rssi) {
         this.rssi = rssi;
     }
 
@@ -52,7 +53,7 @@ public class Device {
         return mtu;
     }
 
-    public void setMtu(Integer mtu) {
+    public void setMtu(@Nullable Integer mtu) {
         this.mtu = mtu;
     }
 
