@@ -1,22 +1,25 @@
 package com.polidea.multiplatformbleadapter;
 
-import android.support.annotation.Nullable;
+
+import androidx.annotation.Nullable;
 
 import java.util.UUID;
 
+/**
+ * @noinspection unused
+ */
 public class ScanResult {
 
     private String deviceId;
     private String deviceName;
     private int rssi;
     private int mtu;
-    @Nullable
     private boolean isConnectable;
     @Nullable
     private UUID[] overflowServiceUUIDs;
     private AdvertisementData advertisementData;
 
-    public ScanResult(String deviceId, String deviceName, int rssi, int mtu, boolean isConnectable, UUID[] overflowServiceUUIDs, AdvertisementData advertisementData) {
+    public ScanResult(String deviceId, String deviceName, int rssi, int mtu, boolean isConnectable, @Nullable UUID[] overflowServiceUUIDs, AdvertisementData advertisementData) {
         this.deviceId = deviceId;
         this.deviceName = deviceName;
         this.rssi = rssi;
@@ -70,7 +73,7 @@ public class ScanResult {
         return overflowServiceUUIDs;
     }
 
-    public void setOverflowServiceUUIDs(UUID[] overflowServiceUUIDs) {
+    public void setOverflowServiceUUIDs(@Nullable UUID[] overflowServiceUUIDs) {
         this.overflowServiceUUIDs = overflowServiceUUIDs;
     }
 
