@@ -1,6 +1,6 @@
 package com.polidea.multiplatformbleadapter.utils;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import com.polidea.multiplatformbleadapter.OnErrorCallback;
 import com.polidea.multiplatformbleadapter.OnSuccessCallback;
@@ -10,9 +10,9 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 public class SafeExecutor<T> {
 
-    private OnSuccessCallback<T> successCallback;
-    private OnErrorCallback errorCallback;
-    private AtomicBoolean wasExecuted = new AtomicBoolean(false);
+    private final OnSuccessCallback<T> successCallback;
+    private final OnErrorCallback errorCallback;
+    private final AtomicBoolean wasExecuted = new AtomicBoolean(false);
 
     public SafeExecutor(@Nullable OnSuccessCallback<T> successCallback, @Nullable OnErrorCallback errorCallback) {
         this.successCallback = successCallback;
@@ -20,13 +20,13 @@ public class SafeExecutor<T> {
     }
 
     public void success(T data) {
-        if (wasExecuted.compareAndSet(false, true)) {
+        if (wasExecuted.compareAndSet(false, true) && successCallback != null) {
             successCallback.onSuccess(data);
         }
     }
 
     public void error(BleError error) {
-        if (wasExecuted.compareAndSet(false, true)) {
+        if (wasExecuted.compareAndSet(false, true) && errorCallback != null) {
             errorCallback.onError(error);
         }
     }
